Implemented configuration parsing into AST
