+++ humphrey-server/src/config/mod.rs
@@ -1,4 +1,7 @@
-mod config;
+pub mod config;
+mod config_old;
+pub mod error;
 pub mod extended_hashmap;
+pub mod traceback;
 
-pub use config::*;
+pub use config_old::*;
