+++ humphrey-server/src/tests/mod.rs
@@ -1,1 +1,2 @@
 pub mod config;
+pub mod config_old;
